refactored laundry service directory orgaization
